Fix request data in Set EDPp Scaling Factor
'''
The Change was required Since Request Data For Set EDPp Scaling
Factor was changed in NSM Spec
'''

Tested on HW

'''

:~# nsmtool telemetry SetEDPpScalingFactors -a 0 -p 0 -s 88 -m 30 -v
nsmtool: <6> Tx: 10 de 80 89 03 0a 03 00 00 58
Success in creating the socket : RC = 5
Success in connecting to socket : RC = 0
Success in sending message type as VDM to mctp demux daemon : RC = 4
nsmtool: <6> Rx: 10 de 00 89 03 0a 00 00 00 00 00
{
    "Completion Code": 0
}

:~# nsmtool telemetry GetEDPpScalingFactors -m 30 -v
nsmtool: <6> Tx: 10 de 80 89 03 09 00
Success in creating the socket : RC = 5
Success in connecting to socket : RC = 0
Success in sending message type as VDM to mctp demux daemon : RC = 4
nsmtool: <6> Rx: 10 de 00 89 03 09 00 00 00 03 00 ff 58 58
{
    "Completion Code": 0,
    "Requested_Persistent_Scaling_Factor": 255,
    "Requested_OneShot_Scaling_Factor": 88,
    "Enforced_Scaling_Factor": 88
}

GET TESTING

curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0/Processors
   /GPU_SXM_1/EnvironmentMetrics
"Oem": {
    "Nvidia": {
      "@odata.type":
           "#NvidiaEnvironmentMetrics.v1_2_0.NvidiaEnvironmentMetrics",
      "EDPpPercent": {
        "@odata.type": "#NvidiaEnvironmentMetrics.v1_0_0.EDPpPercent",
        "AllowableMax": 150,
        "AllowableMin": 81,
        "Persistency": null,
        "SetPoint": 88
      },
      "PowerLimitPersistency": null
    }

PATCH TESTING

Test with Persistency: False
~ # curl -X PATCH 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0
       /Processors/GPU_SXM_1/EnvironmentMetrics
           -d '{"Oem":{"Nvidia":{"EDPpPercent":{"SetPoint":89,
                           "Persistency": false}}}}' -i
HTTP/1.1 200 OK
Allow: GET, PATCH
OData-Version: 4.0
Strict-Transport-Security: max-age=31536000; includeSubdomains
Pragma: no-cache
Cache-Control: no-store, max-age=0
X-Content-Type-Options: nosniff
Date: Tue, 01 Oct 2024 10:56:59 GMT
Content-Length: 0

curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0
               /Processors/GPU_SXM_1/EnvironmentMetrics
"Oem": {
    "Nvidia": {
      "SetPoint" 89;
		 }
		}

**After Driver Reload**

curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0
                        /Processors/GPU_SXM_1/EnvironmentMetrics
    "Nvidia": {
	     "SetPoint": 88;
		 }
		}

Test with Persistency: TRUE

~ # curl -X PATCH 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0
             /Processors/GPU_SXM_1/EnvironmentMetrics
                   -d '{"Oem":{"Nvidia":{"EDPpPercent":{"SetPoint":95,
                             "Persistency": true}}}}' -i
HTTP/1.1 200 OK
Allow: GET, PATCH
OData-Version: 4.0
Strict-Transport-Security: max-age=31536000; includeSubdomains
Pragma: no-cache
Cache-Control: no-store, max-age=0
X-Content-Type-Options: nosniff
Date: Tue, 01 Oct 2024 11:03:03 GMT
Content-Length: 0

curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0/Processors
                            /GPU_SXM_1/EnvironmentMetrics
"Oem": {
    "Nvidia": {
	     "SetPoint": 95;
		 }
		}

**After Driver Unload**

curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0/Processors
                           /GPU_SXM_1/EnvironmentMetrics
"Oem": {
    "Nvidia": {
	     "SetPoint": 255;
		 }
		}
**After driver Reload**

curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0/Processors
                                          /GPU_SXM_1/EnvironmentMetrics
"Oem": {
    "Nvidia": {
	     "SetPoint": 95;
		 }
		}

**POST Reset EDPP percent**

curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0/Processors
                                          /GPU_SXM_1/EnvironmentMetrics
"Oem": {
    "Nvidia": {
	     "SetPoint": 87;
		 }
		}

~ # curl -X POST 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0
                      /Processors/GPU_SXM_1/EnvironmentMetrics/Actions
                               /Oem/NvidiaEnvironmentMetrics.ResetEDPp
{
  "@Message.ExtendedInfo": [
    {
      "@odata.type": "#Message.v1_1_1.Message",
      "Message": "The request completed successfully.",
      "MessageArgs": [],
      "MessageId": "Base.1.18.1.Success",
      "MessageSeverity": "OK",
      "Resolution": "None."
    }
  ]
~

~ # curl -X GET 192.168.31.1/redfish/v1/Systems/HGX_Baseboard_0
                             /Processors/GPU_SXM_1/EnvironmentMetrics
"Oem": {
    "Nvidia": {
	     "SetPoint": 100;
		 }
		}

'''

Fixes nvbug https://nvbugspro.nvidia.com/bug/4850428

+++ nsmtool/nsm_telemetry_cmd.cpp
@@ -2257,7 +2257,7 @@ class SetEDPpScalingFactors : public CommandInterface
   private:
     uint8_t action;
     uint8_t persistence;
-    uint32_t scaling_factor;
+    uint8_t scaling_factor;
 };
 
 class QueryScalarGroupTelemetry : public CommandInterface
